requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
psycopg2-binary
asyncpg